% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas05.R
\name{canvas05}
\alias{canvas05}
\title{Canvas 5}
\usage{
canvas05(
  width = 300,
  height = 500,
  bg_color = "#06060a",
  accent_color = "#c49102",
  planet_color = "#c43235",
  file_name = "canvas05",
  n_circles = 6,
  a0 = 120,
  a1 = 40,
  b0 = 6,
  b1 = 9,
  top = height/3,
  bottom = height/2,
  ...
)
}
\value{

}
\description{
Canvas 5
}
