% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas06.R
\name{canvas06}
\alias{canvas06}
\title{Canvas 6}
\usage{
canvas06(
  width = 300,
  height = 500,
  squareHeight = 50,
  bg_color = "#ffffff",
  accent_color = "#933A16",
  file_name = "canvas06",
  colors = c("#beef9e", "#a6c36f", "#828c51", "#335145", "#1e352f"),
  ...
)
}
\value{

}
\description{
Canvas 6
}
