% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas07.R
\name{canvas07}
\alias{canvas07}
\title{Canvas 7}
\usage{
canvas07(
  width = 300,
  height = 500,
  h = 100,
  bg_color = "#000000",
  colors = c("#6565a4", "#cdedfd", "#b6dcfe", "#a9f8fb", "#31999b"),
  file_name = "canvas07",
  ...
)
}
\value{

}
\description{
Canvas 7
}
