% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-dive.R
\name{canvas08}
\alias{canvas08}
\title{Canvas 8}
\usage{
canvas08(
  width = 300,
  height = 500,
  h = 25,
  bgColour = "#000000",
  colors = c("#31999b", "#cdedfd", "#6565a4", "#b6dcfe", "#a9f8fb"),
  file_name = "08-dive",
  depth = 3,
  r1 = 0.6,
  exp = 8,
  ...
)
}
\value{

}
\description{
Canvas 8
}
