% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-prisma.R
\name{canvas09}
\alias{canvas09}
\title{Canvas 9}
\usage{
canvas09(
  width = 300,
  height = 500,
  bg_color = "#29274C",
  accent_color = "#933A16",
  line_colors = rep(c("#337ca0", "#3ec300", "#fffc31", "#ff1d15", "#e13700"), 50),
  file_name = "09-prisma",
  strips = 15,
  r = 50,
  n = 60,
  ...
)
}
\value{

}
\description{
Canvas 9
}
