% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas01.R
\name{canvas1}
\alias{canvas1}
\title{Canvas 1}
\usage{
canvas1(
  width = 300,
  height = 500,
  squareHeight = 10,
  gutter = 10,
  background_color = "#FDF5E6",
  square_color = "#933A16",
  square_color_shift = "gray40",
  file_name = "canvas1",
  ...
)
}
\value{

}
\description{
Canvas 1
}
