% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-sundial.R
\name{canvas10}
\alias{canvas10}
\title{Canvas 10}
\usage{
canvas10(
  width = 300,
  height = 500,
  bgColour = "#ffffff",
  accent_color = "#933A16",
  file_name = "10-sundial",
  steps = 40,
  ...
)
}
\value{

}
\description{
Canvas 10
}
