% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-breton.R
\name{canvas11}
\alias{canvas11}
\title{Canvas 11}
\usage{
canvas11(
  width = 300,
  height = 500,
  squareHeight = 150,
  lineSize = 1,
  nRows = 6,
  nCols = 4,
  bgColor = "#ffffff",
  lineColor = "#933A16",
  fillColor = "gray90",
  fileName = "11-breton",
  ...
)
}
\value{

}
\description{
Canvas 11
}
