% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-circuit.R
\name{canvas12}
\alias{canvas12}
\title{Canvas 12}
\usage{
canvas12(
  width = 300,
  height = 500,
  bgColour = "#8E6E53",
  accentColour = c("#8E6E53", "#006E53"),
  file_name = "12-circuit",
  ...
)
}
\value{

}
\description{
Canvas 12
}
