% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-klint.R
\name{canvas13}
\alias{canvas13}
\title{Canvas 13}
\usage{
canvas13(
  width = 300,
  height = 500,
  bgColour = "#885053",
  accentColour = c("#777DA7", "#FDFFFC", "#D19C1D", "#2A2B2A", "#8E3B46"),
  fileName = "13-klint",
  ...
)
}
\value{

}
\description{
Canvas 13
}
