% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-meteor.R
\name{canvas15}
\alias{canvas15}
\title{Canvas 15}
\usage{
canvas15(
  width = 300,
  height = 500,
  bgColour = "#2E2E3A",
  accentColours = c("#97ead2", "#8cc7a1", "#816e94", "#74226c", "#4b2142", "#00A6ED"),
  fileName = "15-meteor",
  meteorN = 50,
  meteorWidths = c(50, 80, 120, 150),
  meteorHeights = c(0.5, 1, 1.5, 2),
  meteorRotation = -pi/2.8,
  meteorGradientSplits = 4,
  ...
)
}
\value{

}
\description{
Canvas 15
}
