% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-boxer.R
\name{canvas16}
\alias{canvas16}
\title{Canvas 16}
\usage{
canvas16(
  width = 300,
  height = 500,
  bgColour = "#ffffff",
  accentColours = c("#50514f", "#f25f5c", "#ffe066", "#247ba0", "#70c1b3", "#8e9dcc",
    "#d9dbf1", "#f3d9dc"),
  fileName = "16-boxer",
  colMinWidth = 10,
  colMaxWidth = 50,
  rowMinHeight = 20,
  rowMaxHeight = 60,
  ...
)
}
\value{

}
\description{
Canvas 16
}
