% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas02.R
\name{canvas2}
\alias{canvas2}
\title{Canvas 2}
\usage{
canvas2(
  width = 300,
  height = 500,
  theta = -pi/48,
  scale = 1.15,
  n_steps = 60,
  background_color = "#db451f",
  file_name = "canvas2",
  ...
)
}
\value{

}
\description{
Canvas 2
}
