% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20-melon.R
\name{canvas20}
\alias{canvas20}
\title{Canvas 20}
\usage{
canvas20(
  width = 300,
  height = 500,
  bgColour = "#ffffff",
  accentColours = c("#995d81", "#eb8258", "#f6f740", "#d8dc6a", "#6689a1"),
  fileName = "20-melon",
  n = 20,
  r_min = 10,
  r_max = 30,
  ...
)
}
\value{

}
\description{
Canvas 20
}
