% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25-shortcake.R
\name{canvas25}
\alias{canvas25}
\title{Canvas 25}
\usage{
canvas25(
  width = 300,
  height = 500,
  bgColour = "#F3F9E3",
  accentColours = c("#eacbd2", "#dfaeb4", "#dd9ac2", "#b486ab"),
  fileName = "25-shortcake",
  ...
)
}
\value{

}
\description{
Canvas 25
}
