% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas-template.R
\name{canvas_template}
\alias{canvas_template}
\title{Canvas 2}
\usage{
canvas_template(
  width = 300,
  height = 500,
  background_color = "#FDF5E6",
  square_color = "#933A16",
  file_name = "canvas03",
  ...
)
}
\value{

}
\description{
Canvas 2
}
