% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_quiz.R
\name{export_quiz}
\alias{export_quiz}
\title{Export quiz as TeX}
\usage{
export_quiz(quiz, output_file, class_name = "class name",
  quiz_name = "quiz name", n_ver = 1, shuffle = F)
}
\arguments{
\item{quiz}{a data.frame created by import_quiz}

\item{output_file}{the location to save the TeX document}

\item{class_name}{name of class to be used on quiz}

\item{quiz_name}{name of quiz to be used on quiz}

\item{n_ver}{number of versions to create when TeX is compiled}

\item{shuffle}{Logical indicating whether answer choices should be randomized}
}
\value{
outputs a TeX file to output_file location
}
\description{
Export quiz as TeX
}

