% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quiz}
\alias{quiz}
\title{A dataset containing the prices and other attributes of almost 54,000
diamonds.}
\format{a list created by exporting a quiz from Canvas, then importing into R via:
quiz <- XML::xmlParse("quiz.xml")
quiz <- XML::xmlToList(quiz)}
\source{
Canvas
}
\usage{
quiz
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}

