% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{save_plot}
\alias{save_plot}
\title{Save Plot}
\usage{
save_plot(
  plot,
  file_name,
  width,
  height,
  path = "output",
  echo = TRUE,
  trim = TRUE,
  ...
)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
Save Plot
}
