% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-vertigo.R
\name{canvas02}
\alias{canvas02}
\title{Canvas 2}
\usage{
canvas02(
  width = 300,
  height = 500,
  theta = -pi/48,
  scale = 1.15,
  n_steps = 60,
  background_color = "#db451f",
  file_name = "02-vertigo",
  ...
)
}
\value{

}
\description{
Canvas 2
}
