% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas3.R
\name{canvas03}
\alias{canvas03}
\title{Canvas 3}
\usage{
canvas03(
  width = 300,
  height = 600,
  background_color = "#8D8741",
  line_color = "#FBEEC1",
  file_name = "canvas03",
  ...
)
}
\value{

}
\description{
Canvas 3
}
