% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-lilypad.R
\name{canvas07}
\alias{canvas07}
\title{Canvas 7}
\usage{
canvas07(
  width = 300,
  height = 500,
  squareHeight = 50,
  bgColour = "#ffffff",
  accent_color = "#933A16",
  file_name = "07-lilypad",
  colors = c("#beef9e", "#a6c36f", "#828c51", "#335145", "#1e352f"),
  ...
)
}
\value{

}
\description{
Canvas 7
}
