% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas07.R
\name{canvas07}
\alias{canvas07}
\title{Canvas 7}
\usage{
canvas07(
  width = 300,
  height = 500,
  bg_color = "#ffffff",
  accent_color = "#933A16",
  file_name = "canvas07",
  ...
)
}
\value{

}
\description{
Canvas 7
}
