% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-breton.R
\name{canvas11}
\alias{canvas11}
\title{Canvas 10}
\usage{
canvas11(
  width = 300,
  height = 500,
  bg_color = "#ffffff",
  accent_color = "#933A16",
  file_name = "11-breton",
  ...
)
}
\value{

}
\description{
Canvas 10
}
