% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas-template.R
\name{canvas_template}
\alias{canvas_template}
\title{Canvas _}
\usage{
canvas_template(
  width = 300,
  height = 500,
  bgColour = "#ffffff",
  accentColour = "#933A16",
  fileName = "___",
  ...
)
}
\value{

}
\description{
Canvas _
}
