% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_quiz.R
\name{import_quiz}
\alias{import_quiz}
\title{Import quizzes from xml objects}
\usage{
import_quiz(data, html_patterns = c("<div>", "</div>", "<p>", "</p>"))
}
\arguments{
\item{data}{a list}

\item{html_patterns}{a character vector of patterns to strip from question text}
}
\value{
a data.frame containing questions and answers
}
\description{
Import quizzes from xml objects
}
\examples{
import_quiz(quiz)
}

